Module qml::qtypes

Contains definitions required to communicate with Qt with the use of named types.

Provides definition of QMetaType enum, that contains mapping of named types and a trait QMetaTypable, that controls which types are able to be used in signals, slots or properties.

Enums

QMetaType
Analogue of Qt::QMetaType::Type

Traits

QMetaTypable
Provides a variant of enum for a type.