Struct qml::QVariant

pub struct QVariant {
    // some fields omitted
}

This holds a value to be providen for a QML context.

A value can be different: int, string, float, double, bool or even a custom object.

Methods

impl QVariant

fn to_int(&self) -> i32

fn into_bool(self) -> bool

fn into_float(self) -> f32

fn into_double(self) -> f64

fn set(&mut self, other: &QVariant)

Sets the value for this QVariant