Crate qml

Modules

qtypes
Contains definitions required to communicate with Qt with the use of named types.

Macros

Q_LISTMODEL!
Generates a wrapper for QListModel for static typing and easier management.

Q_OBJECT!
Marks the structure to be able to be used in Qt meta-object system.

qvarlist!
Eases forming of QVariantLists (QVariant of array).

Structs

QListModel
Allows providing a custom model to QML

QVariant
This holds a value to be providen for a QML context.

QmlEngine
Provides an entry point for building QML applications from Rust